tools/llm/auth_analysis_tools.py
"""Authentication and authorization security analysis tools using O1 models."""

from typing import Dict, Any
import logging
import yaml

from .tool_manager import ToolExecutor, ToolResult

class AuthAnalysisTool(ToolExecutor):
    """Tool for analyzing authentication and authorization logic"""

    def __init__(self) -> None:
        self.logger = logging.getLogger(__name__)
        self.model = "gpt-4-1106-preview"

    async def execute(self, **kwargs: Any) -> ToolResult:
        """Execute authentication analysis.

        Args:
            **kwargs: Arguments passed to the tool
                target: The target code or endpoint to analyze
                context: Additional context for the analysis

        Returns:
            ToolResult containing analysis results or error
        """
        try:
            target = str(kwargs.get("target", ""))
            context = kwargs.get("context", {})

            if not target:
                return ToolResult(
                    success=False,
                    error="No target provided for analysis"
                )

            # Parse target if it's YAML
            if target.strip().startswith("{") or target.strip().startswith("---"):
                data = yaml.safe_load(target)
            else:
                data = {"code": target}

            # Add context to data
            data.update(context)

            # Perform analysis
            result = await self._analyze_auth(data)
            return ToolResult(success=True, data=result)

        except Exception as e:
            self.logger.error(f"Auth analysis failed: {str(e)}")
            return ToolResult(success=False, error=str(e))

    async def _analyze_auth(self, data: Dict[str, Any]) -> Dict[str, Any]:
        """Analyze authentication and authorization logic.

        Args:
            data: The parsed data to analyze

        Returns:
            Analysis results
        """
        # Implementation of auth analysis
        return {}
